Get Samozino Profile

get_samozino_profile returns Samozino model estimated from known bodyweight, push_off_distance, mean_GRF_over_distance, mean_velocity and gravity_const.

get_samozino_profile(
  bodyweight,
  push_off_distance,
  mean_GRF_over_distance,
  mean_velocity,
  gravity_const = 9.81
)

Arguments

bodyweight
Numeric value

push_off_distance
Numeric value

mean_GRF_over_distance
Numeric vector

mean_velocity
Numeric vector

gravity_const
Numeric value

Value

Object returned from get_samozino_optimal_profile with addition of RSE and R_Squared elements

References

Samozino, Pierre. ‘A Simple Method for Measuring Lower Limb Force, Velocity and Power Capabilities During Jumping’. In Biomechanics of Training and Testing, edited by Jean-Benoit Morin and Pierre Samozino, 65–96. Cham: Springer International Publishing, 2018. https://doi.org/10.1007/978-3-319-05633-3_4.

———. ‘Optimal Force-Velocity Profile in Ballistic Push-off: Measurement and Relationship with Performance’. In Biomechanics of Training and Testing, edited by Jean-Benoit Morin and Pierre Samozino, 97–119. Cham: Springer International Publishing, 2018. https://doi.org/10.1007/978-3-319-05633-3_5.

Samozino, Pierre, Jean-Benoît Morin, Frédérique Hintzy, and Alain Belli. ‘Jumping Ability: A Theoretical Integrative Approach’. Journal of Theoretical Biology 264, no. 1 (May 2010): 11–18. https://doi.org/10.1016/j.jtbi.2010.01.021.

Samozino, Pierre, Enrico Rejc, Pietro Enrico Di Prampero, Alain Belli, and Jean-Benoît Morin. ‘Optimal Force–Velocity Profile in Ballistic Movements—Altius’: Medicine & Science in Sports & Exercise 44, no. 2 (February 2012): 313–22. https://doi.org/10.1249/MSS.0b013e31822d757a.

Examples

require(tidyverse)

#> Loading required package: tidyverse

#> ── Attaching packages ──────────────────────────────────────────────────────────────────────────────────────────────────────── tidyverse 1.3.0 ──

#> ✓ tibble  3.0.0     ✓ dplyr   0.8.5
#> ✓ tidyr   1.0.2     ✓ stringr 1.4.0
#> ✓ readr   1.3.1     ✓ forcats 0.5.0
#> ✓ purrr   0.3.3     

#> ── Conflicts ─────────────────────────────────────────────────────────────────────────────────────────────────────────── tidyverse_conflicts() ──
#> x dplyr::filter() masks stats::filter()
#> x dplyr::lag()    masks stats::lag()

data("testing_data")

testing_data <- testing_data %>%
  mutate(
    height = vjsim::get_height_from_aerial_time(aerial_time),
    total_load = bodyweight + external_load
  )

jump_metric <- function(data) {
  samozino_metrics <- vjsim::get_samozino_jump_metrics(
    mass = data$bodyweight + data$external_load,
    push_off_distance = data$push_off_distance,
    height = data$height
  )
  return(as.data.frame(samozino_metrics))
}

testing_data <- testing_data %>%
  # Need to add bodyweight so it is kept in the output
  group_by(athlete, bodyweight) %>%
  do(jump_metric(.))

samozino_profile <- function(data) {
  samozino_data <- vjsim::get_samozino_profile(
    bodyweight = data$bodyweight,
    push_off_distance = data$push_off_distance,
    mean_GRF_over_distance = data$mean_GRF_over_distance,
    mean_velocity = data$mean_velocity
  )

  return(as.data.frame(samozino_data))
}

testing_data_samozino <- testing_data %>%
  # Need to add bodyweight so it is kept in the output
  group_by(athlete, bodyweight) %>%
  do(samozino_profile(.))

testing_data_samozino

#> # A tibble: 5 x 29
#> # Groups:   athlete, bodyweight [5]
#>   athlete bodyweight    F0 F0_rel    V0  Pmax Pmax_rel    Sfv Sfv_rel
#>   <fct>        <dbl> <dbl>  <dbl> <dbl> <dbl>    <dbl>  <dbl>   <dbl>
#> 1 John           100 3373.   33.7  2.72 2295.     22.9 -1240.  -12.4 
#> 2 Jack            85 2236.   26.3  4.04 2257.     26.6  -554.   -6.52
#> 3 Peter           95 3772.   39.7  3.49 3295.     34.7 -1079.  -11.4 
#> 4 Jane            55 1961.   35.6  2.31 1133.     20.6  -848.  -15.4 
#> 5 Chris           75   NA    NA   NA      NA      NA      NA    NA   
#> # … with 20 more variables: take_off_velocity <dbl>, height <dbl>,
#> #   optimal_F0 <dbl>, optimal_F0_rel <dbl>, optimal_V0 <dbl>,
#> #   optimal_height <dbl>, optimal_height_diff <dbl>,
#> #   optimal_height_ratio <dbl>, optimal_Pmax <dbl>, optimal_Pmax_rel <dbl>,
#> #   optimal_take_off_velocity <dbl>, optimal_take_off_velocity_diff <dbl>,
#> #   optimal_take_off_velocity_ratio <dbl>, optimal_Sfv <dbl>,
#> #   optimal_Sfv_rel <dbl>, Sfv_perc <dbl>, FV_imbalance <dbl>, probe_IMB <dbl>,
#> #   RSE <dbl>, R_squared <dbl>