vjsim data

Data generated through simulation. Used for exploring relationship between Force Generator characteristics as latent or intrinsic qualities and manifested performance in a single jump as well as profile metrics and optimization metrics

data(vjsim_data)

Format

A data frame with 8640 rows and 174 variables:

force_generator.bodyweight

force_generator.push_off_distance

force_generator.max_force

force_generator.max_velocity

force_generator.decline_rate

force_generator.peak_location

force_generator.time_to_max_activation

force_generator.F0

force_generator.F0_rel

force_generator.V0

force_generator.Pmax

force_generator.Pmax_rel

force_generator.Sfv

force_generator.take_off_velocity

force_generator.height

force_generator.optimal_F0

force_generator.optimal_F0_rel

force_generator.optimal_V0

force_generator.optimal_height

force_generator.optimal_height_diff

force_generator.optimal_height_ratio

force_generator.optimal_Pmax

force_generator.optimal_Pmax_rel

force_generator.optimal_take_off_velocity

force_generator.optimal_take_off_velocity_diff

force_generator.optimal_take_off_velocity_ratio

force_generator.optimal_Sfv

force_generator.Sfv_perc

force_generator.FV_imbalance

force_generator.probe_IMB

bodyweight_jump.take_off_time

bodyweight_jump.take_off_velocity

bodyweight_jump.height

bodyweight_jump.mean_GRF_over_distance

bodyweight_jump.mean_GRF_over_time

bodyweight_jump.mean_velocity

bodyweight_jump.work_done

bodyweight_jump.impulse

bodyweight_jump.mean_power

bodyweight_jump.mean_RFD

bodyweight_jump.mean_RPD

bodyweight_jump.peak_GRF

bodyweight_jump.peak_GRF_time

bodyweight_jump.peak_GRF_distance

bodyweight_jump.peak_velocity

bodyweight_jump.peak_velocity_time

bodyweight_jump.peak_velocity_distance

bodyweight_jump.peak_power

bodyweight_jump.peak_power_distance

bodyweight_jump.peak_power_time

bodyweight_jump.peak_RFD

bodyweight_jump.peak_RFD_distance

bodyweight_jump.peak_RFD_time

bodyweight_jump.peak_RPD

bodyweight_jump.peak_RPD_distance

bodyweight_jump.peak_RPD_time

bodyweight_jump.diff_peak_to_take_off_velocity

bodyweight_jump.ratio_peak_to_take_off_velocity

bodyweight_jump.diff_mean_to_take_off_velocity

bodyweight_jump.ratio_mean_to_take_off_velocity

bodyweight_jump.diff_mean_to_peak_velocity

bodyweight_jump.ratio_mean_to_peak_velocity

bodyweight_jump.diff_mean_to_peak_power

bodyweight_jump.ratio_mean_to_peak_power

bodyweight_jump.diff_mean_GRF_over_distance_to_time

bodyweight_jump.ratio_mean_GRF_over_distance_to_time

bodyweight_jump.mean_velocity_as_TOV_half

probe_bodyweight_jump.height_force

probe_bodyweight_jump.height_force_diff

probe_bodyweight_jump.height_force_ratio

probe_bodyweight_jump.height_velocity

probe_bodyweight_jump.height_velocity_diff

probe_bodyweight_jump.height_velocity_ratio

probe_bodyweight_jump.height_diff

probe_bodyweight_jump.height_ratio

probe_bodyweight_jump.take_off_velocity_force

probe_bodyweight_jump.take_off_velocity_force_diff

probe_bodyweight_jump.take_off_velocity_force_ratio

probe_bodyweight_jump.take_off_velocity_velocity

probe_bodyweight_jump.take_off_velocity_velocity_diff

probe_bodyweight_jump.take_off_velocity_velocity_ratio

probe_bodyweight_jump.take_off_velocity_diff

probe_bodyweight_jump.take_off_velocity_ratio

bosco.height_2BW

bosco.index

profile_mean_FV.F0

profile_mean_FV.F0_rel

profile_mean_FV.V0

profile_mean_FV.Pmax

profile_mean_FV.Pmax_rel

profile_mean_FV.Sfv

profile_mean_power.Pmax

profile_mean_power.Pmax_rel

profile_mean_power.Pmax_location

profile_mean_power.F0_perc

profile_peak_FV.F0

profile_peak_FV.F0_rel

profile_peak_FV.V0

profile_peak_FV.Pmax

profile_peak_FV.Pmax_rel

profile_peak_FV.Sfv

profile_peak_power.Pmax

profile_peak_power.Pmax_rel

profile_peak_power.Pmax_location

profile_peak_power.F0_perc

profile_load_take_off_velocity.V0

profile_load_take_off_velocity.L0

profile_load_take_off_velocity.L0_rel

profile_load_take_off_velocity.Imax

profile_load_take_off_velocity.Imax_rel

profile_load_take_off_velocity.Slv

profile_load_impulse.Imax

profile_load_impulse.Imax_rel

profile_load_impulse.Imax_location

profile_load_impulse.L0_perc

samozino_theoretical_profile.F0

samozino_theoretical_profile.F0_rel

samozino_theoretical_profile.V0

samozino_theoretical_profile.Pmax

samozino_theoretical_profile.Pmax_rel

samozino_theoretical_profile.Sfv

samozino_theoretical_profile.take_off_velocity

samozino_theoretical_profile.height

samozino_theoretical_profile.optimal_F0

samozino_theoretical_profile.optimal_F0_rel

samozino_theoretical_profile.optimal_V0

samozino_theoretical_profile.optimal_height

samozino_theoretical_profile.optimal_height_diff

samozino_theoretical_profile.optimal_height_ratio

samozino_theoretical_profile.optimal_Pmax

samozino_theoretical_profile.optimal_Pmax_rel

samozino_theoretical_profile.optimal_take_off_velocity

samozino_theoretical_profile.optimal_take_off_velocity_diff

samozino_theoretical_profile.optimal_take_off_velocity_ratio

samozino_theoretical_profile.optimal_Sfv

samozino_theoretical_profile.Sfv_perc

samozino_theoretical_profile.FV_imbalance

samozino_theoretical_profile.probe_IMB

samozino_practical_profile.F0

samozino_practical_profile.F0_rel

samozino_practical_profile.V0

samozino_practical_profile.Pmax

samozino_practical_profile.Pmax_rel

samozino_practical_profile.Sfv

samozino_practical_profile.take_off_velocity

samozino_practical_profile.height

samozino_practical_profile.optimal_F0

samozino_practical_profile.optimal_F0_rel

samozino_practical_profile.optimal_V0

samozino_practical_profile.optimal_height

samozino_practical_profile.optimal_height_diff

samozino_practical_profile.optimal_height_ratio

samozino_practical_profile.optimal_Pmax

samozino_practical_profile.optimal_Pmax_rel

samozino_practical_profile.optimal_take_off_velocity

samozino_practical_profile.optimal_take_off_velocity_diff

samozino_practical_profile.optimal_take_off_velocity_ratio

samozino_practical_profile.optimal_Sfv

samozino_practical_profile.Sfv_perc

samozino_practical_profile.FV_imbalance

samozino_practical_profile.probe_IMB

simple_profile.L0

simple_profile.TOV0

simple_profile.Sfv

simple_profile.take_off_velocity

simple_profile.height

simple_profile.optimal_L0

simple_profile.optimal_TOV0

simple_profile.optimal_Sfv

simple_profile.optimal_take_off_velocity

simple_profile.optimal_height

simple_profile.Sfv_perc

simple_profile.FV_imbalance

simple_profile.probe_IMB